% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mevr.R
\name{event_separation}
\alias{event_separation}
\title{Separates rainfall timeseries by dry time events}
\usage{
event_separation(
  d,
  separation_in_min = 360,
  time_resolution = 10,
  ignore_event_duration = 30,
  min_rain = 0.1
)
}
\arguments{
\item{d}{Data as tibble with observation date and values.}

\item{separation_in_min}{Dry minutes between ordinary events (storms). 
Defaults to 360 minutes, i.e. 6 hours (depends on local climate). 
Values less than \code{min_rain} count as zero.}

\item{time_resolution}{e.g. 10 minutes}

\item{ignore_event_duration}{Events are ignored when their duration is shorter than 
ignore_event_duration minutes. Defaults to 30 minutes.}

\item{min_rain}{Rainfall below min_rain is ignored.}
}
\value{
Returns a list with two components. The data series with an 
additional column is_event useful in the next step of the event separation. 
And a tibble with indices from, to defining the events in the input timeseries.
}
\description{
Based on Marra...
}
\details{
min_rain could also be calculated form the timeseries (see Marra, et al., 2020)
}
