% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mevr.R
\name{print.mevr}
\alias{print.mevr}
\title{Print method for object of class mevr}
\usage{
\method{print}{mevr}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Object of class \code{mevr}, fitted with MEVD, SMEV or TMEV.}

\item{digits}{Number of digits.}

\item{...}{Additional parameters.}
}
\value{
A nicely formatted output of the fitting results.
}
\description{
Print nicely formatted output of the fit to the MEVD and its variants
}
\examples{
data(dailyrainfall)

# fit a simplified MEVD
fit <- fsmev(dailyrainfall)
print(fit)
}
