% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mevr.R
\name{ordinary_events}
\alias{ordinary_events}
\title{Identifies ordinary rainfall events by calculating the maximum within rainfall events
defined with \code{\link{event_separation}}}
\usage{
ordinary_events(data, duration)
}
\arguments{
\item{data}{A list with tibbles of the input data where the events are defined,
and indices of the single events. Usually the output of function \code{\link{event_separation}}.}

\item{duration}{The duration in minutes for which maxima shall be calculated.}
}
\value{
Returns a tibble with individual rainfall events that can be 
used as input for functions \code{\link{fsmev}}, \code{\link{fmev}}, \code{\link{ftmev}}.
}
\description{
Identifies ordinary rainfall events by calculating the maximum within rainfall events
defined with \code{\link{event_separation}}
}
